LED-85 refactor: streamline imports in user persistence mapper and repository for improved clarity
